refactor(admin y newProduct): modularizacion con custom hooks y pages
